backend/requirements.txt
openai==0.28.0
numpy
pdfplumber>=0.9.0
python-docx>=0.8.11
tiktoken>=0.4.0
faiss-cpu>=1.7.4
httpx==0.28.1
beautifulsoup4>=4.12.2